requirements.txt
asgiref
Django
django-ckeditor
django-colorfield
django-cors-headers
django-js-asset
djangorestframework
pillow
psycopg
psycopg2-binary
pytz
sqlparse
typing_extensions
tzdata
certifi
charset-normalizer
django-parler
django-parler-rest
django-rosetta
idna
polib
requests
urllib3